feat(backend): Drop unused checkCrossOrigin util
